pyproject.toml
[project]
name = "notesplus"
version = "0.1.0"
description = "Notes application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "apscheduler>=3.11.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "passlib>=1.7.4",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.40",
    "types-python-jose>=3.4.0.20250224",
]

[build-system]
requires = ["setuptools==75.8.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
lint = [
    "mypy>=1.15.0",
    "ruff>=0.11.8",
    "types-passlib>=1.7.7.20250408",
]
